+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
   * Add onlyObjectOrNil for sets.
 
+  * Add method to extract the first object passing test from an array.
+
 2.5.1, 22-July-2014
 
   * Fix isSubsetOf to correctly compare mutable/immutable variants.